feat: renamed the directory and added an installation script
