+++ src/examples/test_default_in_spellbook.crl
