+++ src/examples/test_oop_one_self.crl
