+++ src/examples/test_oop_two_self.crl
