+++ src/main.go
@@ -3,10 +3,9 @@ package main
 import (
 	"fmt"
 	"os"
-
-	"thecarrionlanguage/evaluator"
-	"thecarrionlanguage/object"
-	"thecarrionlanguage/repl"
+	"thecarrionlanguage/src/evaluator"
+	"thecarrionlanguage/src/object"
+	"thecarrionlanguage/src/repl"
 )
 
 const CROW_IMAGE = `
@@ -32,7 +31,7 @@ func main() {
 	env := object.NewEnvironment()
 
 	// Attempt to load the standard library from "munin/" folder
-	err := evaluator.LoadMuninStdlib(env, "munin")
+	err := evaluator.LoadMuninStdlib(env, "src/munin")
 	if err != nil {
 		fmt.Printf("Error loading Munin standard library: %v\n", err)
 		// You might decide to exit here or continue without the stdlib
