+++ src/object/bound_method.go
