+++ src/object/object_test.go
